fix: apply physical offset correction and timing optimization for reflexive sensor
- Lowered physical offset from 29.5cm to 25.8cm (reduction of 3.7cm)
- Reduced timeout_us from 25000 to 13000 for faster recovery
- Increased ping_interval_ms from 70 to 100
- Increased PING_COUNT from 7 to 9
- Updated TankGeometry constructor and member to float
- Re-added temporary individual ping logs
- Bumped version to 0.2.2 for OTA

+++ main/include/tank_geometry.hpp
@@ -35,7 +35,7 @@ public:
      *                         level (e.g. overflow drain). This is the only
      *                         runtime calibration parameter.
      */
-    explicit TankGeometry(uint8_t sensor_offset_cm)
+    explicit TankGeometry(float sensor_offset_cm)
         : offset_cm_(sensor_offset_cm)
     {
     }
@@ -106,7 +106,7 @@ private:
 
     static constexpr uint8_t LUT_SIZE = sizeof(VOLUME_LUT) / sizeof(VOLUME_LUT[0]);
 
-    const uint8_t offset_cm_; ///< Sensor distance (cm) when tank is at full level.
+    const float offset_cm_; ///< Sensor distance (cm) when tank is at full level.
 
     /**
      * @brief Interpolate volume permille from a depth value using the LUT.
